add PriceToWin

+++ MeliLibToolsNext/APIs/Request/Products/Struct.cs
@@ -10,7 +10,7 @@ namespace MeliLibToolsNext.APIs.Request.Products
     public class Struct
     {
         [JsonProperty("number")] 
-        public int? Number { get; set; }
+        public double? Number { get; set; }
 
         [JsonProperty("unit")]
         public string Unit { get; set; }
