+++ MeliLibToolsNext/APIs/Request/User/UserRequest.cs
@@ -7,7 +7,7 @@ using MeliLibToolsNext.Attributes;
 
 namespace MeliLibToolsNext.APIs.Request.User
 {
-    public class UserRequest(int? userId = null) : RequestBase(query:
+    public class UserRequest(long? userId = null) : RequestBase(query:
     [
         new(name: "User_id", parameterType: QueryParameterType.VALUE_ONLY,order:1,value:userId),
     ]);
